Verschieben des Reports nach "gegen Gast"
